Refactor weapon application and remove BLE task
- Removed the BLE task implementation as it is no longer needed.
- Updated control task to handle trigger button initialization and shooting logic.
- Introduced display manager for handling UI updates and state management.
- Added runtime metrics for system monitoring.
- Refactored game task to simplify game state handling.
- Updated laser task to use a 32-bit message format.
- Implemented ESP-NOW communication for hit events and game state broadcasting.
- Created debug print functionality for NVS contents.
- Added WiFi task for managing WiFi connections.
- Updated display task to remove dependency on LVGL and integrate with the new display manager.

+++ src/tasks/laser_task.cpp
@@ -10,7 +10,7 @@ static const char* TAG = "LaserTask";
 
 extern QueueHandle_t laserMessageQueue;
 
-static void sendMessage(uint16_t message)
+static void sendMessage(uint32_t message)
 {
     TickType_t nextWake = xTaskGetTickCount();
     const TickType_t bitPeriod = pdMS_TO_TICKS(BIT_DURATION_MS);
@@ -27,7 +27,7 @@ static void sendMessage(uint16_t message)
 void laser_task(void* pvParameters)
 {
     ESP_LOGI(TAG, "Laser task started");
-    uint16_t message;
+    uint32_t message;
 
     while (1)
     {
